Only cast EmacsInt to ptrdiff_t/isize at the very end
Before this I was doing things the other way around, casting f
and t to isize several times in the function body. For some reason
this caused a compare-strings test to fail, because despite
receiving 'foo as the value of "from" it would not signal a
wrong argument type error.